+++ src/ArchInspec/Command/InspectCommand.php
@@ -18,13 +18,14 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 /**
  * This command is the entry point for evaluating the architecture based on defined architecture definitions.
  */
 class InspectCommand extends Command
 {
-    const EXIT_SUCCESS = 0, EXIT_VIOLATION = 1;
+    const EXIT_SUCCESS = 0, EXIT_VIOLATION = 1, EXIT_FAILED = 3;
 
     const ARGUMENT_CONFIG = 'config';
     const OPTION_REPORT_UNDEFINED = 'reportUndefined';
@@ -34,7 +35,7 @@ class InspectCommand extends Command
      */
     protected function configure()
     {
-        $defaultConfig = __DIR__ . '/../../../archinspec.yml.dist';
+        $defaultConfig = realpath(__DIR__ . '/../../../archinspec.yml.dist');
         $this->addArgument(self::ARGUMENT_CONFIG, InputArgument::OPTIONAL, CliMessage::ARGUMENT_CONFIG, $defaultConfig);
         $this->addOption(self::OPTION_REPORT_UNDEFINED, 'u', InputOption::VALUE_NONE,
             CliMessage::OPTION_REPORT_UNDEFINED);
@@ -46,14 +47,21 @@ class InspectCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $configFile = $input->getArgument(self::ARGUMENT_CONFIG);
-
         $output->writeln($this->getDescription() . PHP_EOL);
-        $output->writeln(CliMessage::READ_CONFIG_FROM . $configFile . PHP_EOL);
 
+        // read config
+        $output->writeln(CliMessage::READ_CONFIG_FROM . $configFile);
         $config = AIConfig::fromYamlFile($configFile);
         if ($input->getOption(self::OPTION_REPORT_UNDEFINED)) {
             $config->setReportUndefined(true);
         }
+
+        if (!$this->isValidArchitectureFile($config, $input, $output)) {
+            return self::EXIT_FAILED;
+        }
+        $output->writeln(CliMessage::READ_ARCHITECTURE_FILE_FROM . $config->getArchitecture());
+        $output->writeln(CliMessage::STARTING_ANALYSIS);
+        $output->writeln("");
 
         $archInspec = new ArchInspec($config);
         if ($archInspec->analyze(new ConsoleWriter($output))) {
@@ -61,6 +69,40 @@ class InspectCommand extends Command
         } else {
             return self::EXIT_VIOLATION;
         }
+    }
+
+    /**
+     * Makes sure the architecture file is readable and helps the user if not.
+     *
+     * @param AIConfig $config
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     *
+     * @return bool
+     */
+    private function isValidArchitectureFile(AIConfig $config, InputInterface $input, OutputInterface $output)
+    {
+        $arcFile = $config->getArchitecture();
+        if (is_readable($arcFile)) {
+            return true;
+        }
+
+        if (file_exists($arcFile)) {
+            // permission error
+            $output->writeln(sprintf(CliMessage::ARCHITECTURE_FILE_NOT_READABLE, $arcFile));
+
+        } else {
+            // no architecture file present
+            $helper = $this->getHelper('question');
+            $question = new ConfirmationQuestion(sprintf(CliMessage::ARCHITECTURE_FILE_ASK_TO_CREATE, $arcFile), false);
+            if ($helper->ask($input, $output, $question)) {
+                touch($arcFile);
+                $output->writeln(sprintf(CliMessage::ARCHITECTURE_FILE_CREATED_EMPTY, $arcFile));
+            }
+            $output->writeln(CliMessage::ARCHITECTURE_FILE_HOWTO);
+        }
+
+        return false;
     }